+++ src/main/java/io/nats/client/Nats.java
@@ -72,7 +72,7 @@ public class Nats {
     /**
      * Current version of the library - {@value #CLIENT_VERSION}
      */
-    public static final String CLIENT_VERSION = "2.5.0";
+    public static final String CLIENT_VERSION = "2.5.1";
 
     /**
      * Current language of the library - {@value #CLIENT_LANGUAGE}